projeto2025/projeto2025/backend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
